Add files via upload
Updated grant proposal for Jamaica infrastructure in support of WeatherXM and other sensors based on feedback from IoT Working Group. 

This update includes additional sensors for soil, air quality, tracking and safety that (a) are complimentary to the weather data for the farmers where the stations will be deployed, and (b) will drive more traffic on the IoT network. It also includes detail on the cost of off-grid hotspots (which are needed in more agricultural and other remote monitoring systems), and includes costs for video and other marketing collateral production.